src/API_LAB/requirements.txt
fastapi
uvicorn
pymysql
python-dotenv
cryptography
email-validator
